fixed comments step2
