+++ src/services/reducers/index.js
@@ -1,14 +1,14 @@
 import { combineReducers } from "redux";
 
-import ingredientReducer from "./reducers";
 import constructorReducer from "./constructorReducer";
 import ingredientsReducer from "./ingredientsReducer";
 import ingredientDetailsReducer from "./ingredientDetailsReducer";
+import orderReducer from "./orderReducer";
 
 const rootReducer = combineReducers({
   constructorIngredients: constructorReducer,
   currentIngredient: ingredientDetailsReducer,
-  ingredientReducer,
+  order: orderReducer,
   ingredients: ingredientsReducer,
 });
