+++ src/services/reducers/orderReducer.js
@@ -1,0 +1,42 @@
+import {
+  GET_ORDER_NUMBER_REQUEST,
+  GET_ORDER_NUMBER_SUCCESS,
+  GET_ORDER_NUMBER_FAILED,
+} from "../actions/orderAction";
+
+const orderInitialState = {
+  orderNumber: 0,
+  orderNumberRequest: false,
+  orderNumberFailed: false,
+};
+
+const orderReducer = (state = orderInitialState, action) => {
+  switch (action.type) {
+    case GET_ORDER_NUMBER_REQUEST: {
+      return {
+        ...state,
+        orderNumberRequest: true,
+      };
+    }
+    case GET_ORDER_NUMBER_SUCCESS: {
+      return {
+        ...state,
+        orderNumberRequest: false,
+        orderNumberFailed: false,
+        orderNumber: action.number,
+      };
+    }
+    case GET_ORDER_NUMBER_FAILED: {
+      return {
+        ...state,
+        orderNumberRequest: false,
+        orderNumberFailed: true,
+      };
+    }
+    default: {
+      return state;
+    }
+  }
+};
+
+export default orderReducer;
